vintern/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
transformers
Pillow
numpy
einops
timm
